move playwright fixtures to separate module
